Created QuoteTest.java for JUnitTesting
